cli_dev/requirements.txt
click==8.1.3
colorama==0.4.6
PyPDF2==3.0.1
